Split binary_search into integer and floating.
