+++ crates/algolib/binary_search_floating/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "binary_search_floating"
+version = "0.1.0"
+authors = ["Kana <ngtkana@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+
+[dev-dependencies]
+assert-impl = "0.1.3"
+rand = "0.8.4"
+randtools = { path = "../../utils/randtools/" }
