+++ crates/algolib/binary_search_integer/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "binary_search"
+name = "binary_search_integer"
 version = "0.1.0"
 authors = ["Kana <ngtkana@gmail.com>"]
 edition = "2018"
